small fix on memory

+++ lib/src/data/memory/long_term_memory.dart
@@ -3,6 +3,7 @@ part of 'meal_db_adapter.dart';
 class _LongTermMemory {
   final MealDataBase longTermMemory = MealDataBase(boxName: 'clientBox');
   static const _longTermMemoryDuration = Duration(hours: 8);
+  // static const _longTermMemoryDuration = Duration(seconds: 1);
 
   Future<void> clearLongTermMemoryInternal() async {
     await longTermMemory.clear();
@@ -16,15 +17,28 @@ class _LongTermMemory {
         if (isValidLongTermMemory(cacheData.creationDate)) {
           return cacheData.value;
         } else {
-          debugPrint("[readLongTermMemory]>> $key expired on long memory ");
-          await deleteLongTermMemory(key);
-          return;
+          if (key.toString().toLowerCase().startsWith('clientkeys')) {
+            debugPrint(
+              "[readLongTermMemory]>> return $key even expired, because is config value",
+            );
+            return cacheData.value;
+          } else {
+            debugPrint(
+              "[readLongTermMemory]>> $key is not config and expired on long memory removing... "
+                  .toUpperCase(),
+            );
+            await deleteLongTermMemory(key);
+            return;
+          }
         }
       }
-      debugPrint("[readLongTermMemory]>> $key not found on long memory ");
+      debugPrint(
+        "[readLongTermMemory]>> $key not found on long memory".toUpperCase(),
+      );
       return;
     } catch (e) {
-      debugPrint("[readLongTermMemory]>> $key cant access memory $e");
+      debugPrint(
+          "[readLongTermMemory]>> $key cant access memory $e".toUpperCase());
       return;
     }
   }
@@ -45,7 +59,6 @@ class _LongTermMemory {
   bool isValidLongTermMemory(DateTime creation) {
     final currentTime = DateTime.now();
     final difference = currentTime.difference(creation);
-    debugPrint("diffe>> ${difference <= _longTermMemoryDuration}");
     return difference <= _longTermMemoryDuration;
   }
 }
